apps/users/requirements.txt
# requirements.txt
# Semua package di-pin ke versi spesifik untuk reproducible builds
# dan mencegah supply chain attack via auto-upgrade ke versi breaking/vulnerable.
#
# Update procedure:
#   1. pip install pip-tools
#   2. Edit requirements.in dengan constraint versi
#   3. pip-compile requirements.in --output-file requirements.txt
#   4. Jalankan test suite
#   5. Commit requirements.txt
#
# Security audit: pip install pip-audit && pip-audit -r requirements.txt

# ─── Core Framework ───────────────────────────────────────────────────────────
# Migrations project menggunakan Django 6.0.1 — pin ke minor version
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0

# ─── Security & Auth ─────────────────────────────────────────────────────────
django-axes==7.0.1
django-csp==3.8
fido2==1.1.3
pyotp==2.9.0
cryptography==44.0.2          # Untuk Fernet (TOTP secret encryption)

# ─── Database ─────────────────────────────────────────────────────────────────
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# ─── WebSocket & Async ────────────────────────────────────────────────────────
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# ─── Task Queue ───────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.2.1

# ─── Static Files & Server ───────────────────────────────────────────────────
whitenoise==6.9.0
gunicorn==23.0.0
uvicorn==0.34.0

# ─── Social Auth / OAuth ─────────────────────────────────────────────────────
social-auth-app-django==5.4.3
social-auth-core==4.5.4

# ─── Audit / History ─────────────────────────────────────────────────────────
django-simple-history==3.8.0

# ─── Email ────────────────────────────────────────────────────────────────────
sendgrid==6.11.0

# ─── QR Code / Image ─────────────────────────────────────────────────────────
qrcode==8.1
pillow==12.1.1

# ─── API Docs ─────────────────────────────────────────────────────────────────
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.4.1
django-filter==24.3

# ─── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
